Program operational.
Ended up doing some hacky work, creating a database and a user
just to be able to use the ardrive.getByMetaDataTxFromSyncTable -function.

The code now successfully retrieves data about public files on
the given Drive ID under an Arweave-address, containing three modes
for CSV-output. Had to write the data into a file as arweave-core-js
was putting garbage into stdout.

Made this quite impulsively within a short time as the CSV-export
was a feature requested by many on the ArDrive Discord server.
It was a good learning experience, being actually the first time
I write anything in JavaScript. It isn't too different than my
current main language, Java, so adapting to it was no trouble.